+++ src/main/java/com/progra3/proyecto/service/IVehiculoService.java
@@ -8,10 +8,10 @@ import com.progra3.proyecto.entity.Vehiculo;
 public interface IVehiculoService {
 	
     public List<Vehiculo> getAll();
-    public Vehiculo getById(Integer id);
+    public Vehiculo getById(Long id);
     public Vehiculo save(Vehiculo vehiculo);
-    public void delete(Integer id);
-    public boolean exists(Integer id);
+    public void delete(Long id);
+    public boolean exists(Long id);
     
     public List<Vehiculo> findByTipo (String tipo);
